requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
Django==2.1.7
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.3
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.5
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tox==3.14.5
urllib3==1.25.8
virtualenv==20.0.8
wcwidth==0.1.8
zipp==3.1.0
